🐛 change statement order in _filterGithubData()
To make 'No Data found' work on comments checking for i=0 and null data
first is required. If statements with elif and else only evaluate one
and so the more specific statement needs to come first. Also I forgot to
reset noData to be empty so it can be reused.